requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-paypal==2.0
paypalrestsdk==1.13.3
pytest==8.3.2
pytest-django==4.8.0
pytest-cov
coverage